fix: example
