+++ example/sample-fetch.ts
@@ -1,27 +1,11 @@
-import * as Formatter from "@himenon/openapi-parameter-formatter";
 import fetch from "node-fetch";
 
 import { ApiClient, Client, HttpMethod, ObjectLike, QueryParameters } from "./client";
+import { generateQueryString } from "./utils";
 
 export interface RequestOption {
   timeout?: number;
 }
-
-const generateQueryString = (queryParameters: QueryParameters | undefined): string | undefined => {
-  if (!queryParameters) {
-    return undefined;
-  }
-  return Object.entries(queryParameters).reduce((queryString, [key, item]) => {
-    if (!item.style) {
-      return queryString + "&" + `${key}=${item}`;
-    }
-    const result = Formatter.QueryParameter.generate(key, item as Formatter.QueryParameter.Parameter);
-    if (result) {
-      return queryString + "&" + result;
-    }
-    return queryString;
-  }, "");
-};
 
 const apiClientImpl: ApiClient<RequestOption> = {
   request: async (
@@ -33,9 +17,9 @@ const apiClientImpl: ApiClient<RequestOption> = {
     options?: RequestOption,
   ): Promise<any> => {
     const query = generateQueryString(queryParameters);
-    const requestUrl = query ? url + "&" + query : url;
+    const requestUrl = query ? url + "?" + encodeURI(query) : url;
     const response = await fetch(requestUrl, {
-      body: requestBody,
+      body: JSON.stringify(requestBody),
       headers,
       method: httpMethod,
       timeout: options?.timeout,
